- Changed ony's animation names. - Pillow logic implemented (Different attacks, collisions, input response). Now Ony can execute attacks and combo chains. - Created classes to show a text display next to a game object (Not used yet).

+++ trunk/OUAN/OUAN/Src/Logic/LogicComponent/LogicComponentEnemy.h
@@ -62,7 +62,7 @@ namespace OUAN
 		int mColourSensitivityMask;	
 
 		/// 'Immunity' time so a single collision is not processed during several ticks
-		int mHitRecoveryTime;
+		double mHitRecoveryTime;
 
 		std::string getColourName(int colour);
 		std::string getMaskString();
